fixed compiler warnings; added debug statements

+++ include/block.hpp
@@ -33,18 +33,18 @@ class Block {
     }
 
     void InsertInstructions(std::vector<instr_ptr> instrs, size_t idx) {
-        instructions.insert(instructions.begin() + idx,
+        instructions.insert(instructions.begin() + static_cast<long>(idx),
                             std::make_move_iterator(instrs.begin()),
                             std::make_move_iterator(instrs.end()));
     }
 
     void InsertInstruction(instr_ptr instr, size_t idx) {
-        instructions.insert(instructions.begin() + idx, std::move(instr));
+        instructions.insert(instructions.begin() + static_cast<long>(idx), std::move(instr));
     }
 
     void RemoveInstruction(size_t idx) {
         assert(idx < instructions.size());
-        instructions.erase(instructions.begin() + idx);
+        instructions.erase(instructions.begin() + static_cast<long>(idx));
     }
 
     std::vector<instr_ptr> ReleaseInstructions() {
@@ -60,7 +60,7 @@ class Block {
 
     void RemoveSuccessor(size_t idx) {
         assert(idx < successors.size());
-        successors.erase(successors.begin() + idx);
+        successors.erase(successors.begin() + static_cast<long>(idx));
     }
 
     void AddPredecessor(Block *pred) { predecessors.push_back(pred); }
@@ -72,7 +72,7 @@ class Block {
 
     void RemovePredecessor(size_t idx) {
         assert(idx < predecessors.size());
-        predecessors.erase(predecessors.begin() + idx);
+        predecessors.erase(predecessors.begin() + static_cast<long>(idx));
     }
 
     InstructionBase *GetInstruction(size_t idx) {
